requirements.txt
# The following are the dependencies for a Python project that involves reinforcement learning using the gymnasium and stable-baselines3 libraries, along with PyTorch, NumPy, and Matplotlib:
gymnasium[classic-control]>=0.29.1
stable-baselines3>=2.3.0
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
moviepy==1.0.3
imageio[ffmpeg]>=2.36.0
scipy>=1.10.0
tensorboard>=2.14.0
pyyaml>=6.0.0
